create navbar component
