+++ src/App.tsx
@@ -1,30 +1,10 @@
-import { NavLink, Outlet } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import './App.scss';
-import classNames from 'classnames';
-
-const getLinkClass = ({ isActive }: { isActive: boolean }) =>
-  classNames('navbar-item', { 'has-background-grey-lighter': isActive });
+import { Navbar } from './components/Navbar/Navbar';
 
 export const App = () => (
   <div data-cy="app">
-    <nav
-      data-cy="nav"
-      className="navbar is-fixed-top has-shadow"
-      role="navigation"
-      aria-label="main navigation"
-    >
-      <div className="container">
-        <div className="navbar-brand">
-          <NavLink className={getLinkClass} to="/">
-            Home
-          </NavLink>
-
-          <NavLink className={getLinkClass} to="/people">
-            People
-          </NavLink>
-        </div>
-      </div>
-    </nav>
+    <Navbar />
 
     <main className="section">
       <div className="container">
